Paper + smaating

+++ project2/main.cpp
@@ -5,7 +5,7 @@
 #include "matrixmul.h"
 #include "test.h"
 
-typedef Matrix<RowBased<int32_t>, Naive> M0;
+typedef Matrix<RowBased<double>, Naive> M0;
 typedef Matrix<ColumnBased<double>, Naive> M1;
 
 using namespace std;
